% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Threshold.R
\name{calc_Threshold}
\alias{calc_Threshold}
\title{calc_Threshold.}
\usage{
calc_Threshold(myG, p = 0.05)
}
\arguments{
\item{myG}{Genotype file}

\item{p}{desired p value}
}
\value{
GWAS threshold
}
\description{
calculate thresholds.
}
