% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTaxaNames.R
\name{checkNYCvsKansas}
\alias{checkNYCvsKansas}
\title{checkNYCvsKansas}
\usage{
checkNYCvsKansas(folder = "GWAS_Results/", deleteKansas = F)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{deteleKansas}{Logical, if TRUE, will delete any \code{Kansas} files with no difference between the \code{NYC} files.}
}
\value{
A table indicating the number of markers with different p values between NYC and Kansas files.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_NYCvsKansas.html}{Check if NYC and Kansas results are identical, if so, the Kansas file will be deleted.}
}
