% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNYCvsKansas.R
\name{checkTaxaNames}
\alias{checkTaxaNames}
\title{checkTaxaNames}
\usage{
checkTaxaNames(xG = myG, xY = myY)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format and read in with header = F.}

\item{xY}{GWAS phenotype object.}
}
\value{
A list of names not present in both datasets.
}
\description{
Check to see if the names in your myY and myG file match.
}
