% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_GWAS_Results.R
\name{clean_GWAS_Results}
\alias{clean_GWAS_Results}
\title{clean_GWAS_Results}
\usage{
clean_GWAS_Results(
  folder = "GWAS_Results/",
  remove_nonResults = F,
  remove_Kansas = F,
  remove_NYC = F
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{remove_nonResults}{Delelets all non .csv files (all .pdf files)}

\item{remove_Kansas}{Delete Kansas files.}

\item{remove_NYC}{Delete NYC files.}
}
\value{
A table of which models have been run on each trait.
}
\description{
Delete extra GWAS files in the folder containing GWAS result files from GAPIT.
}
