% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_IUPAC.R
\name{convert_IUPAC}
\alias{convert_IUPAC}
\title{convert_IUPAC}
\usage{
convert_IUPAC(xg, toIUPAC = T)
}
\arguments{
\item{toIUPAC}{Logical, convert to IUPAC (TRUE) or from IUPAC (FALSE).}

\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}
}
\value{
A LD Heatmap plot.
}
\description{
Converts hapmap formatted genotype files to and from IUPAC codes.
}
