% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_plotly.R
\name{gg_GWAS_plotly}
\alias{gg_GWAS_plotly}
\title{gg_GWAS_plotly}
\usage{
gg_GWAS_plotly(mp, filename = "GWAS_plotly.html")
}
\arguments{
\item{mp}{ggplot object.}

\item{filename}{Filename for plot.}
}
\value{
A plotly summary GWAS plot.
}
\description{
Creates an interactive plotly plot.
}
