% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_Summary.R
\name{gg_GWAS_Summary}
\alias{gg_GWAS_Summary}
\title{gg_GWAS_Summary}
\usage{
gg_GWAS_Summary(
  folder = "GWAS_Results/",
  traits = list_Traits(folder),
  groups = NULL,
  threshold = round(-log10(5e-08), 1),
  sug.threshold = round(-log10(5e-06), 1),
  chr = NULL,
  pos1 = NULL,
  pos2 = NULL,
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "CMLM", "SUPER"),
  model.colors = gwaspr_Colors,
  shapes = 21:25,
  hlines = NULL,
  vlines = NULL,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  threshold.legend = T,
  title = "Summary of Significant GWAS Results",
  caption = NULL,
  rowread = 2000,
  legend.position = "bottom",
  legend.rows = 1,
  chr.unit = "100 Mbp",
  plotHBPvalues = F,
  skyline = "NYC"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{groups}{Grouping for the traits. Should be equal length to \code{traits}.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{chr}{Chromosomes to plot.}

\item{pos1}{starting position to plot.}

\item{pos2}{ending position to plot.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model.}

\item{shapes}{The shape values to use for the different models. e.g., 21:25.}

\item{hlines}{Locations for horizontal lines. e.g., hlines = c(1.5,2.5).}

\item{vlines}{Markers to be labelled with a vertical red line.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, display of vline color legend.}

\item{threshold.legend}{Logical, display of vline color legend.}

\item{title}{A title for the plot.}

\item{caption}{A caption for the plot.}

\item{rowread}{Number of rows to read for each GWAS results file.}

\item{legend.rows}{Number of rows for the legend.}

\item{chr.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{plotHBPvalues}{Logical, should H.B.P.Values be uses.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}
}
\value{
A GWAS summary plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_GWAS_Summary.html}{Creates a summary GWAS plot of significant associations.}
Note: this function requires the GWAS results files to be ordered
}
