% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_Summary_plotly.R
\name{gg_GWAS_Summary_plotly}
\alias{gg_GWAS_Summary_plotly}
\title{gg_GWAS_Summary_plotly}
\usage{
gg_GWAS_Summary_plotly(
  mp,
  filename = "GWAS_Summary.html",
  width = 10,
  height = 8
)
}
\arguments{
\item{mp}{Plot from gg_GWAS_Summary.}

\item{filename}{Filename for plot.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\value{
A plotly summary GWAS plot.
}
\description{
Creates a summary GWAS plot of significant associations.
}
