% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_LD_Decay.R
\name{gg_LD_Decay}
\alias{gg_LD_Decay}
\title{gg_LD_Decay}
\usage{
gg_LD_Decay(xG = myG, outputFolder, markerNum = 1000)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{outputFolder}{Folder to place RData files into.}

\item{markerNum}{Number of markers to randomly select for each chromosome.}
}
\value{
LD decay plots.
}
\description{
Creates a number of LD decay plots.
}
