% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_LDHeatmap.R
\name{gg_LDheatmap}
\alias{gg_LDheatmap}
\title{gg_LDHeatmap}
\usage{
gg_LDheatmap(
  xG = myG,
  chr = 6,
  pos1 = 0,
  pos2 = 6e+06,
  myMs = NULL,
  myTitle = NULL,
  axisTextSize = NULL,
  nameTrim = NULL
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{chr}{Chromosome to plot.}

\item{pos1}{Start position within the selected chromosome.}

\item{pos2}{End position within the selected chromosome.}

\item{myMs}{Markers to highlight within the plot.}

\item{myTitle}{Title for the plot.}

\item{axisTextSize}{Text size for the axis labels (genotype names).}

\item{nameTrim}{String used to trim marker names.}
}
\value{
A LD Heatmap plot.
}
\description{
Creates a manhattan plot.
}
