% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manh.R
\name{gg_Manh}
\alias{gg_Manh}
\title{gg_Manh}
\usage{
gg_Manh(
  folder = NULL,
  trait = NULL,
  subtitle = NULL,
  markers = NULL,
  labels = markers,
  lines = F,
  facet = T,
  models = c("GLM", "MLM", "CMLM", "MLMM", "SUPER", "FarmCPU", "Blink"),
  colors1 = c("darkgreen", "darkgoldenrod3", "darkgreen", "darkgoldenrod3",
    "darkgreen", "darkgoldenrod3", "darkgreen"),
  colors2 = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4",
    "darkgoldenrod2")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{subtitle}{A subtitle for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{lines}{Logical value of whether or not to include vertical lines with markers.}

\item{models}{Models to read.}

\item{colors1}{Colors for each chromosome}

\item{colors2}{Colors for each model}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot.
}
