% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manh_Zoom.R
\name{gg_Manh_Zoom}
\alias{gg_Manh_Zoom}
\title{gg_Manh_Zoom}
\usage{
gg_Manh_Zoom(
  folder = "Results_Add/",
  trait = "Testa_Pattern",
  chr = 1,
  start = 1e+05,
  end = 1e+06,
  markers = NULL,
  labels = markers,
  lines = F,
  models = c("GLM", "MLM", "CMLM", "MLMM", "SUPER", "FarmCPU", "Blink"),
  colors = c("darkgreen", "darkgoldenrod3", "darkgreen", "darkgoldenrod3", "darkgreen",
    "darkgoldenrod3", "darkgreen")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{lines}{Logical value of whether or not to include vertical lines with markers.}

\item{models}{Models to read.}

\item{colors}{Colors for each chromosome}

\item{subtitle}{A subtitle for the plot.}
}
\value{
A manhattan plot.
}
\description{
Removes columns without any values.
}
