% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan.R
\name{gg_Manhattan}
\alias{gg_Manhattan}
\title{gg_Manhattan}
\usage{
gg_Manhattan(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  chr = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  legend = T,
  legend.rows = 1,
  legend.box = "horizontal",
  point.sizes = c(0.3, 1, 0.75),
  pmax = NULL,
  pmin = 0,
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM", "CMLM", "SUPER"),
  model.colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4",
    "blue2", "magenta3"),
  facet = F,
  highlight.sig = F,
  sig.color = "black",
  chr.colors = rep(c("darkgreen", "darkgoldenrod3"), 30),
  chr.unit = "100 Mbp",
  plotHBPvalues = F,
  skyline = "NYC",
  addQQ = T
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{chr}{Chromosomes to plot. Use if you want to plot a single chromosome.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{legend}{Logical, whether or not to add a legend.}

\item{legend.rows}{Number of rows for the legend.}

\item{legend.box}{Alignment of the legend. Default is "horizontal", but it can be changed to "vertical".}

\item{point.sizes}{Sizes for the points. c("Not Sig", "Sig", "Sug").}

\item{pmax}{A max value for the y-axis. Markers with higher values will be lowered to pmax..}

\item{pmin}{A min Value for plotting. Markers with lower values will be removed.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model. Used if \code{facet = F}.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot. Default is \code{facet = F}.}

\item{highlight.sig}{Logical, whether or not to highlight significant associations with a black circle. Used if \code{facet = F}.}

\item{sig.color}{Color for significant assoctiations. Used if \code{facet = T}.}

\item{chr.colors}{Colors for each chromosome. Used if \code{facet = T}.}

\item{chr.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{plotHBPvalues}{Logical, if TRUE, H.B.P.values be uses.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}

\item{addQQ}{Logical, whether or not to add a QQ plot.}
}
\value{
A manhattan plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Manhattan.html}{Create manhattan plots from GAPIT GWAS results.}
}
