% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan_Traits.R
\name{gg_Manhattan_Traits}
\alias{gg_Manhattan_Traits}
\title{gg_Manhattan_Traits}
\usage{
gg_Manhattan_Traits(
  folder = "GWAS_Results/",
  traits = list_Traits(folder)[1:2],
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = F,
  markers = NULL,
  labels = markers,
  facet = F,
  addQQ = T,
  pmax = NULL,
  model = "MLM",
  trait.colors = c("darkgreen", "darkorange3", "steelblue", "darkred", "darkorchid4",
    "burlywood4", "darkseagreen4"),
  chrom.unit = "100 Mbp",
  legend.rows = 1
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot.}

\item{addQQ}{Logical, whether or not to add a QQ plot}

\item{pmax}{A max value for the y-axis.}

\item{model}{Model to read.}

\item{trait.colors}{Colors for each trait.}

\item{chrom.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{legend.rows}{Number of rows for the legend.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot.
}
