% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan_Zoom1.R
\name{gg_Manhattan_Zoom1}
\alias{gg_Manhattan_Zoom1}
\title{gg_Manhattan_Zoom1}
\usage{
gg_Manhattan_Zoom1(
  folder = "GWAS_Results/",
  traits = list_Traits(folder)[1],
  title = NULL,
  chrom,
  pos1,
  pos2,
  threshold = NULL,
  sug.threshold = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  pmax = NULL,
  model = "MLM",
  model.colors = gwaspr_Colors,
  facet = F,
  highlight.sig = F,
  sig.color = "red",
  legend.rows = 1,
  plotHBPvalues = F,
  skyline = "Kansas"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{title}{A title for the plot.}

\item{chrom}{Chromosome to plot.}

\item{pos1}{Start position on chromosome.}

\item{pos2}{End position on chromosome.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{pmax}{A max value for the y-axis.}

\item{model}{Model to read.}

\item{model.colors}{Colors for each model. Used if `facet = F`.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot. Default is `facet = F`.}

\item{highlight.sig}{Logical, whether or not to highlight significant associations with a black circle. Used if `facet = F`.}

\item{sig.color}{Color for significant assoctiations.}

\item{legend.rows}{Number of rows for the legend.}

\item{plotHBPvalues}{Logical, should H.B.P.Values be uses.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If left NULL, it will use the highest P.value.}

\item{addQQ}{Logical, whether or not to add a QQ plot}

\item{chrom.colors}{Colors for each chromosome. Used if `facet = T`.}

\item{chrom.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot zoomed in to a particular region.
}
