% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan_Zoom_Traits.R
\name{gg_Manhattan_Zoom_Traits}
\alias{gg_Manhattan_Zoom_Traits}
\title{gg_Manhattan_Zoom_Traits}
\usage{
gg_Manhattan_Zoom_Traits(
  folder = "GWAS_Results/",
  traits = list_Traits(folder)[1],
  title = NULL,
  chr,
  pos1 = NULL,
  pos2 = NULL,
  threshold = NULL,
  sug.threshold = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  pmax = NULL,
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM", "CMLM", "SUPER"),
  model.colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4",
    "blue2", "magenta3"),
  sig.color = "black",
  legend.rows = 1,
  legend.box = "horizontal",
  point.sizes = c(0.3, 1, 0.75),
  plotHBPvalues = F,
  skyline = "NYC"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{title}{A title for the plot.}

\item{chr}{Chromosome to plot.}

\item{pos1}{Start position on chromosome.}

\item{pos2}{End position on chromosome.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model.}

\item{sig.color}{Color for significant assoctiations.}

\item{legend.rows}{Number of rows for the legend.}

\item{legend.box}{Alignment of the legend. Default is "horizontal", but it can be changed to "vertical".}

\item{point.sizes}{Sizes for the points. c("Not Sig", "Sig", "Sug").}

\item{plotHBPvalues}{Logical, should H.B.P.Values be uses.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}
}
\value{
A manhattan plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Manhattan_Zoom_Traits.html}{Create manhattan plots for multiple traits zoomed in to a particular region.}
}
