% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker.R
\name{gg_Marker}
\alias{gg_Marker}
\title{gg_Manh}
\usage{
gg_Marker(
  trait,
  marker,
  marker2 = NULL,
  type = "bar",
  points = T,
  colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4", "darkslategray",
    "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{trait}{.}
}
\value{
Plot.
}
\description{
Removes columns without any values.
}
