% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Bubble.R
\name{gg_Marker_Bubble}
\alias{gg_Marker_Bubble}
\title{gg_Marker_Bubble}
\usage{
gg_Marker_Bubble(
  xG,
  xY,
  trait.x = NULL,
  trait.y = "Alleles",
  facet.rows = NULL,
  facet.cols = NULL,
  markers,
  marker.colors = gwaspr_Colors,
  remove.hets = F,
  remove.N = T,
  legend.pos = "bottom",
  legend.rows = 1,
  title = NULL,
  subtitle = paste(markers, collapse = "\\n")
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{trait.x}{Trait for the x-axis.}

\item{trait.y}{Trait for the y-axis.}

\item{facet.rows}{Trait for the row facetting.}

\item{facet.cols}{Trait for the col facetting.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Colors to fill the marker bubbles.}

\item{remove.hets}{Logical, Whether to remove hets or not. advised if plotting multiple markers.}

\item{remove.N}{Logical, Whether to remove N or NN.}

\item{legend.pos}{Position for the legend.}

\item{legend.rows}{Number of rows for the legend.}

\item{title}{Title for the plot.}

\item{subtitle}{Subtitle for the plot. Defaults to the list of markers.}
}
\value{
Marker plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Marker_Bubble.html}{Creates a marker scatter plot with myG and myY objects.}
}
