% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Pie.R
\name{gg_Marker_Pie}
\alias{gg_Marker_Pie}
\title{gg_Marker_Pie}
\usage{
gg_Marker_Pie(
  xG,
  xY,
  trait,
  trait.label = trait,
  trait.levels = NULL,
  markers,
  marker.colors = gwaspr_Colors,
  remove.hets = F,
  remove.N = T,
  title = NULL,
  subtitle = paste(markers, collapse = "\\n"),
  ncol = NULL,
  legend.rows = 1,
  removeHets = T,
  groupByTrait = F
)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{trait.label}{Label for the trait.}

\item{trait.levels}{Factor levels for the trait.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}

\item{remove.hets}{Logical, Whether to remove hets or not. advised if plotting multiple markers.}

\item{remove.N}{Logical, Whether to remove N or NN.}

\item{title}{Title for the plot.}

\item{subtitle}{Subtitle for the plot. Defaults to the list of markers.}

\item{ncol}{number of columns for facetting.}

\item{legend.rows}{number of rows in legend.}

\item{groupByTrait}{Logical, if TRUE, will make pies of each trait instead of each marker.}
}
\value{
Marker plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Marker_Pie.html}{Creates a marker plot with myG and myY objects.}
}
