% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Scatter.R
\name{gg_Marker_Scatter}
\alias{gg_Marker_Scatter}
\title{gg_Marker_Scatter}
\usage{
gg_Marker_Scatter(
  xG,
  xY,
  trait.x,
  trait.y,
  markers,
  marker.colors = gwaspr_Colors,
  remove.hets = T,
  point.size = 1,
  title = NULL,
  legend.rows = 1,
  subtitle = paste(markers, collapse = "\\n")
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{trait.x}{Trait for the x-axis.}

\item{trait.y}{Trait for the y-axis.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Colors to fill in the violin and boxplots.}

\item{remove.hets}{Logical, Whether to remove hets or not. advised if plotting multiple markers.}

\item{point.size}{Size for the points.}

\item{title}{Title for the plot.}

\item{legend.rows}{Number of rows for the legend.}

\item{subtitle}{Subtitle for the plot. Defaults to the list of markers.}

\item{yLab}{Label for the y-axis.}
}
\value{
Marker plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Marker.html}{Creates a marker scatter plot with myG and myY objects.}
}
