% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_myG_Details.R
\name{gg_myG_Details}
\alias{gg_myG_Details}
\title{gg_myG_Details}
\usage{
gg_myG_Details(filename, myPrefix = "", filetype = "csv")
}
\arguments{
\item{filename}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{myPrefix}{Prefix for file names.}

\item{filetype}{default is "csv". other option is "txt".}
}
\value{
2 .csv files with marker details & 3 Marker plots.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_myG_Details.html}{Analyses your genotype data and outputs 2 summary files with marker details and 3 plots.}
}
