% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotly.R
\name{gg_plotly}
\alias{gg_plotly}
\title{gg_plotly}
\usage{
gg_plotly(mp, filename = "my_gg_plotly.html", showlegend = T)
}
\arguments{
\item{mp}{ggplot object.}

\item{filename}{Filename for plot.}

\item{showlegend}{Logical, should the legend be displayed.}
}
\value{
A plotly summary GWAS plot.
}
\description{
Creates an interactive plotly plot. Note: This function will require `plotly` and `htmlwidgets` packages.
}
