% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_QTL_Summary.R
\name{gg_QTL_Summary}
\alias{gg_QTL_Summary}
\title{gg_QTL_Summary}
\usage{
gg_QTL_Summary(
  myG,
  myQ,
  title = "Summary of QTL Results",
  lodFill = F,
  fillColor = "darkgreen",
  fillColor_low = "grey50",
  xLab = "Pos",
  facetLab = "lg"
)
}
\arguments{
\item{myG}{Genetic map. Must contain the following 3 columns c("Marker","Chr","Pos").}

\item{myQ}{Table of QTL results. Must contain the following 5 columns c("Trait","Chr","Pos","Pos_lo","Pos_hi","lod").}

\item{title}{Custom title for the plot.}

\item{lodFill}{Logical. If true, fill color will be a gradient based on lod values.}

\item{fillColor}{Color for filling points.}

\item{fillColor_low}{If \code{lodfill = T}, This will be the low end of the gradeint color scale for filling points.}

\item{xLab}{Custom x lab.}

\item{facetLab}{Custom label for facets.}
}
\value{
A QTL summary plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_QTL_Summary.html}{Creates a summary QTL plot of significant associations.}
}
