% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Vol.R
\name{gg_Vol}
\alias{gg_Vol}
\title{gg_Vol}
\usage{
gg_Vol(
  folder,
  trait,
  subtitle = NULL,
  markers = NULL,
  labels = markers,
  models = c("GLM", "MLM", "MLMM", "CMLM", "SUPER", "FarmCPU", "Blink")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{subtitle}{A subtitle for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{models}{Models to read.}
}
\value{
A volcano plot.
}
\description{
Creates a volcano plot.
}
