% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Volcano.R
\name{gg_Volcano}
\alias{gg_Volcano}
\title{gg_Volcano}
\usage{
gg_Volcano(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  title = trait,
  markers = NULL,
  labels = markers,
  models = c("MLM", "FarmCPU", "BLINK", "MLMM", "GLM", "CMLM", "SUPER"),
  threshold = NULL,
  skyline = "NYC"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{models}{Models to read.}

\item{threshold}{Significant Threshold.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}
}
\value{
A volcano plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Volvano.html}{Creates a volcano plot.}
}
