% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Ran.R
\name{is_Ran}
\alias{is_Ran}
\title{is_Ran}
\usage{
is_Ran(folder = "GWAS_Results/", myY)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{myY}{Phenotype file with traits as column names which will be checked for GWAS results.}
}
\value{
A table of which models have been run on each trait.
}
\description{
Create a table of which models have been run on each trait.
}
