% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_GWAS_Results.R
\name{list_GWAS_Results}
\alias{list_GWAS_Results}
\title{list_GWAS_Results}
\usage{
list_GWAS_Results(
  folder,
  file,
  threshold = 6.7,
  threshold2 = NULL,
  g.range = 3e+06,
  rowread = 2000
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{file}{The file to read.}

\item{threshold}{Significant threshold.}

\item{threshold2}{Suggestive threshold.}

\item{g.range}{Range around peak association.}

\item{rowread}{Number of rows to read.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results.
}
