% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Result_Files.R
\name{list_Result_Files}
\alias{list_Result_Files}
\title{list_Result_Files}
\usage{
list_Result_Files(folder)
}
\arguments{
\item{folder}{Folder containing GWAS results.}
}
\value{
List of GWAS results.
}
\description{
Get list of GWAS results.
}
