% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Results.R
\name{list_Results}
\alias{list_Results}
\title{list_Results}
\usage{
list_Results(folder = "C:/gitfolder/gwas_tutorial/Results", fullfilename = F)
}
\arguments{
\item{folder}{Folder containing GWAS results.}
}
\value{
List of GWAS results.
}
\description{
Get list of GWAS results.
}
