% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Threshold.R
\name{list_Threshold}
\alias{list_Threshold}
\title{list_Threshold.}
\usage{
list_Threshold(num_markers, p = 0.05)
}
\arguments{
\item{num_markers}{Number of markers}

\item{p}{desired p value}
}
\value{
GWAS threshold
}
\description{
calculate thresholds.
}
