% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Top_Markers.R
\name{list_Top_Markers}
\alias{list_Top_Markers}
\title{list_Top_Markers}
\usage{
list_Top_Markers(
  folder = "GWAS_Results/",
  traits = list_Traits(folder),
  chr = 1:50,
  models = c("MLM", "FarmCPU", "BLINK", "MLMM", "GLM", "CMLM", "SUPER"),
  threshold = 5,
  n = 5
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{GWAS trait.}

\item{chr}{Chromosomes to include.}

\item{models}{GWAS models to include.}

\item{threshold}{filters results with -log10(p) below threshold.}

\item{n}{Number per trait and model.}
}
\value{
Table of top results.
}
\description{
Finds the markers with the highest association on each chromosome.
}
