% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Traits.R
\name{list_Traits}
\alias{list_Traits}
\title{list_Traits}
\usage{
list_Traits(folder = "GWAS_Results/")
}
\arguments{
\item{folder}{Folder containing GWAS results.}
}
\value{
List of traits with GWAS results.
}
\description{
Get a list of traits with GWAS results.
}
