% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{movingAverage.}
\usage{
movingAverage(x, n = 3, centered = FALSE)
}
\arguments{
\item{x}{The vector}

\item{n}{The number of samples}

\item{centered}{If FALSE, then average current sample and previous (n-1) samples. If TRUE, then average symmetrically in past and future. (If n is even, use one more sample from future.) the colnames they should be changed to}
}
\value{
data with fixed colnames
}
\description{
This function calculates a moving average.
}
