% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.R
\name{mv}
\alias{mv}
\title{mv}
\usage{
mv(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{The factor or vector to modify.}

\item{from}{A vector of the items to replace.}

\item{to}{A vector of replacement values.}

\item{warn_missing}{print a message if any of the old values are not actually present in x.}
}
\value{
plyr::mapvalues output
}
\description{
A copy of the plyr::mapvalues function.
}
