% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_gg_Manhattan.R
\name{old_gg_Manhattan}
\alias{old_gg_Manhattan}
\title{old_gg_Manhattan}
\usage{
old_gg_Manhattan(
  folder,
  trait,
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.legend = F,
  facet = T,
  qq = T,
  pmax = NULL,
  models = c("MLM", "MLMM", "FarmCPU", "Blink", "GLM"),
  chrom.colors = c("darkgreen", "darkgoldenrod3", "darkgreen", "darkgoldenrod3",
    "darkgreen", "darkgoldenrod3", "darkgreen"),
  model.colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4",
    "darkgoldenrod2")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.legend}{Logical, wheterh or not to add a legend for the vlines.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot.}

\item{qq}{Logical, whether or not to add a QQ plot}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{chrom.colors}{Colors for each chromosome. Used if `facet = T`.}

\item{model.colors}{Colors for each model. Used if `facet = F`.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot. Note: this function is meant for the old version of GAPIT.
}
