% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_gg_Manhattan_Zoom.R
\name{old_gg_Manhattan_Zoom}
\alias{old_gg_Manhattan_Zoom}
\title{old_gg_Manhattan_Zoom}
\usage{
old_gg_Manhattan_Zoom(
  folder,
  trait,
  chr,
  start,
  end,
  title = trait,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.color = "red",
  models = c("GLM", "MLM", "CMLM", "MLMM", "SUPER", "FarmCPU", "Blink"),
  colors = c("darkgreen", "darkgoldenrod3", "darkgreen", "darkgoldenrod3", "darkgreen",
    "darkgoldenrod3", "darkgreen")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{chr}{Chromosome to plot.}

\item{start}{Start position on chromosome.}

\item{end}{End position on chromosome.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.color}{color for each vertical line.}

\item{models}{Models to read.}

\item{colors}{Colors for each chromosome}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot zoomed in to a particular region. Note: this function is meant for the old version of GAPIT.
}
