% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_gg_Volcano.R
\name{old_gg_Volcano}
\alias{old_gg_Volcano}
\title{old_gg_Volcano}
\usage{
old_gg_Volcano(
  folder,
  trait,
  title = trait,
  markers = NULL,
  labels = markers,
  models = c("GLM", "MLM", "MLMM", "CMLM", "SUPER", "FarmCPU", "Blink")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{models}{Models to read.}
}
\value{
A volcano plot.
}
\description{
Creates a volcano plot. Note: this function is meant for the old version of GAPIT.
}
