% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_table_GWAS_Results.R
\name{old_table_GWAS_Results}
\alias{old_table_GWAS_Results}
\title{old_table_GWAS_Results}
\usage{
old_table_GWAS_Results(
  folder,
  files,
  threshold = 6,
  sug.threshold = NULL,
  rowread = 2000
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{files}{The files to read.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{rowread}{Number of rows to read.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results. Note: this function is meant for the old version of GAPIT.
}
