% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_GWAS_Results.R
\name{order_GWAS_Results}
\alias{order_GWAS_Results}
\title{order_GWAS_Results}
\usage{
order_GWAS_Results(folder, files)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{files}{The files to read.}
}
\value{
Rearranged files of GWAS results.
}
\description{
Read in GWAS results, then arrange by P.value and write as csv..
}
