% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GWAS_Region.R
\name{table_GWAS_Region}
\alias{table_GWAS_Region}
\title{table_GWAS_Region}
\usage{
table_GWAS_Region(
  folder = "GWAS_Results/",
  chr = 1,
  pos1 = 3.5e+08,
  pos2 = 3.7e+08,
  fnames = list_Result_Files(folder),
  threshold = 6,
  sug.threshold = NULL,
  nrowstoread = 1000,
  skyline = "NYC",
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM", "CMLM", "SUPER")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{chr}{Chromosome to plot.}

\item{pos1}{Start position on chromosome.}

\item{pos2}{End position on chromosome.}

\item{fnames}{The files to read.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{nrowstoread}{Number of rows to read.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}

\item{models}{GWAS models to use.}

\item{useHBPvalues}{Logical, if TRUE, H.B.P.Values will be uses.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results.
}
