% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GWAS_Results_Summary.R
\name{table_GWAS_Results_Summary}
\alias{table_GWAS_Results_Summary}
\title{table_GWAS_Results_Summary}
\usage{
table_GWAS_Results_Summary(xx, onlySig = F, binMarkers = F, binSize = 1e+06)
}
\arguments{
\item{xx}{Object from `table_GWAS_Results()`.}

\item{onlySig}{Logical, if TRUE, any suggested associations will be removed.}

\item{binMarkers}{Logical, if TRUE, markers will be bined based on `binSize`.}

\item{binSize}{range on each side of marker to bin. default = 1,000,000.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a summary using the output from `table_GWAS_Results()`.
}
