% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_LD_Decay.R
\name{table_LD_Decay}
\alias{table_LD_Decay}
\title{table_LD_Decay}
\usage{
table_LD_Decay(xG = myG, outputFolder)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{outputFolder}{Folder to place RData files into.}
}
\value{
RData files in the outputFolder location of LD data.
}
\description{
Calculates pairwise LD for markers within your myG genotype file.
}
