% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_Results_Summary.R
\name{table_Results_Summary}
\alias{table_Results_Summary}
\title{table_Results_Summary}
\usage{
table_Results_Summary(folder = "GWAS_Results/", isOrdered = F)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{isOrdered}{set to True to see if results files are ordered based on P.value}
}
\value{
A table of which models have been run on each trait.
}
\description{
Create a table of which models have been run on each trait.
}
