% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gwaspr2.R
\name{theme_gwaspr2}
\alias{theme_gwaspr2}
\title{theme_gwaspr2.}
\usage{
theme_gwaspr2(
  x,
  bgFill = "grey95",
  lineColor = "white",
  lineSize = 0.75,
  stripFill = "white",
  caption.adj = 1,
  caption.size = 7,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{lineSize}{border line size}

\item{stripFill}{strip background color}

\item{caption.adj}{hjust for the caption}

\item{caption.size}{size of the caption}
}
\value{
ggplot with gwaspr theme
}
\description{
ggplot theme.
}
