% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gwaspr_col.R
\name{theme_gwaspr_col}
\alias{theme_gwaspr_col}
\title{theme_gwaspr_col.}
\usage{
theme_gwaspr_col(
  x,
  bgFill = "white",
  lineColor = "grey90",
  lineSize = 0.75,
  stripFill = "white",
  caption.adj = 1,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{lineSize}{border line size}

\item{stripFill}{strip background color}

\item{caption.adj}{hjust for the caption}
}
\value{
ggplot with gwaspr theme
}
\description{
ggplot theme.
}
