% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gwaspr_Pie.R
\name{theme_gwaspr_pie}
\alias{theme_gwaspr_pie}
\title{theme_gwaspr_pie.}
\usage{
theme_gwaspr_pie(
  x,
  bgFill = "white",
  lineSize = 0.75,
  stripFill = "white",
  caption.adj = 1,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineSize}{border line size}

\item{stripFill}{strip background color}

\item{caption.adj}{hjust for the caption}
}
\value{
ggplot with agData theme for pie graphs
}
\description{
ggplot theme.
}
