% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_Markers.R
\name{top_Markers}
\alias{top_Markers}
\title{top_Markers}
\usage{
top_Markers(x, chroms = 1:7, n = 1, threshold_filter = T)
}
\arguments{
\item{x}{data.}

\item{chroms}{Chromosomes to include.}

\item{n}{Number per chromosome.}

\item{threshold_filter}{Logical, filters non-significant results.}
}
\value{
Table of top results.
}
\description{
Finds the markers with the highest association on each chromosome.
}
