% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LD_Decay.R
\name{calc_LD_Decay}
\alias{calc_LD_Decay}
\title{calc_LD_Decay}
\usage{
calc_LD_Decay(xG = myG, outputFolder, markerNum = 1000)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{outputFolder}{Folder to place RData files into.}

\item{markerNum}{Number of markers to randomly select for each chromosome.}
}
\value{
RData files in the outputFolder location of LD data.
}
\description{
Calculates pairwise LD for markers within your myG genotype file.
}
