% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNYCvsKansas.R
\name{checkNYCvsKansas}
\alias{checkNYCvsKansas}
\title{checkNYCvsKansas}
\usage{
checkNYCvsKansas(folder = "GWAS_Results/", deleteKansas = F)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{deteleKansas}{Logical, if TRUE, will delete any `Kansas` files with no difference between the `NYC` files.}
}
\value{
A table of which runs were identical and deleted.
}
\description{
Check if NYC and Kansas results are identical, if so, the Kansas file will be deleted.
}
