% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_Hits.R
\name{gg_GWAS_Hits}
\alias{gg_GWAS_Hits}
\title{gg_GWAS_Hits}
\usage{
gg_GWAS_Hits(
  xx,
  xG,
  xCV = NULL,
  traits,
  range = 2e+06,
  title = "",
  sigMin = 0,
  models = c("MLM", "FarmCPU", "BLINK", "MLMM", "GLM", "CMLM", "SUPER"),
  model.colors = c("darkgreen", "darkorange3", "steelblue", "darkred", "darkorchid4",
    "burlywood4", "darkseagreen4"),
  model.shapes = c(21:25, 21:22),
  vlines = NULL,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  legend.rows = 1
)
}
\arguments{
\item{xx}{Table of significant GWAS results. See ?table_GWAS_Results().}

\item{xG}{Genotype data in hapmap format.}

\item{xCV}{Optional, for filtering if you have a "CV" column.}

\item{traits}{List of traits to use.}

\item{range}{Range for binning GWAS hits.}

\item{title}{Title for horizontal facet.}

\item{sigMin}{Minimum number of hits to plot.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model.}

\item{vlines}{Markers to be labelled with a vertical red line.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{legend.rows}{Number of rows for the legend.}
}
\value{
A GWAS Hits plot.
}
\description{
Creates a summary GWAS plot of significant associations.
}
