% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_Summary.R
\name{gg_GWAS_Summary}
\alias{gg_GWAS_Summary}
\title{gg_GWAS_Summary}
\usage{
gg_GWAS_Summary(
  folder,
  files,
  threshold,
  threshold2 = NULL,
  markers = NULL,
  hlines = F,
  g.range = 3e+06,
  rowread = 2000,
  plotly = F,
  plotly_filename = "GWAS_Summary.html",
  caption = paste0("Threshold = ", threshold, "\\nSuggestive = ", threshold2)
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{markers}{Markers to be labelled.}

\item{trait}{The trait to read.}

\item{subtitle}{A subtitle for the plot.}

\item{labels}{Labels to be used for markers.}

\item{lines}{Logical value of whether or not to include vertical lines with markers.}

\item{models}{Models to read.}

\item{colors1}{Colors for each chromosome}

\item{colors2}{Colors for each model}
}
\value{
A manhattan plot.
}
\description{
Creates a summary GWAS plot of significant associations.
}
