% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GWAS_Summary.R
\name{gg_GWAS_Summary}
\alias{gg_GWAS_Summary}
\title{gg_GWAS_Summary}
\usage{
gg_GWAS_Summary(
  folder = NULL,
  traits = list_Traits(),
  threshold = -log10(5e-08),
  threshold2 = -log10(5e-06),
  models = c("MLM", "MLMM", "FarmCPU", "Blink", "GLM"),
  colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkgoldenrod2"),
  shapes = 21:25,
  hlines = NULL,
  markers = NULL,
  markers2 = NULL,
  markers3 = NULL,
  title = NULL,
  caption = paste0("Sig Threshold = ", threshold, " = Large\\nSuggestive = ",
    threshold2, " = Small"),
  rowread = 2000
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{threshold}{Significant threshold.}

\item{threshold2}{Suggestive threshold.}

\item{models}{Models to read.}

\item{colors}{Colors for each model.}

\item{hlines}{Locations for horizontal lines. e.g., hlines = c(1.5,2.5).}

\item{markers}{Markers to be labelled with a vertical red line.}

\item{title}{A title for the plot.}

\item{caption}{A caption for the plot.}

\item{rowread}{Number of rows to read for each GWAS results file.}
}
\value{
A GWAS summary plot.
}
\description{
Creates a summary GWAS plot of significant associations.
}
