% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_LD_Heatmap.R
\name{gg_LD_Heatmap}
\alias{gg_LD_Heatmap}
\title{gg_LD_Heatmap}
\usage{
gg_LD_Heatmap(
  xG = myG,
  chr,
  pos1,
  pos2,
  metric = "R^2",
  threshold = 0.9,
  myMs = NULL,
  myTitle = NULL,
  axisTextSize = NULL,
  nameTrim = NULL,
  color.low = "white",
  color.mid = "goldenrod1",
  color.high = "darkred"
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{chr}{Chromosome to plot.}

\item{pos1}{Start position within the selected chromosome.}

\item{pos2}{End position within the selected chromosome.}

\item{metric}{Which LD calculation to use. Default is "D'".}

\item{threshold}{Value for selecting linked markers. default is "0.9".}

\item{myMs}{Markers to highlight within the plot.}

\item{myTitle}{Title for the plot.}

\item{axisTextSize}{Text size for the axis labels (genotype names).}

\item{nameTrim}{String used to trim marker names.}

\item{color.low}{Color for gradient low.}

\item{color.mid}{Color for gradient mid.}

\item{color.high}{Color for gradient high.}
}
\value{
A LD Heatmap plot.
}
\description{
Creates a LD heatmap plot.
}
