% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan.R
\name{gg_Manhattan}
\alias{gg_Manhattan}
\title{gg_Manhattan}
\usage{
gg_Manhattan(
  folder,
  trait,
  title = trait,
  threshold = NULL,
  threshold2 = NULL,
  markers = NULL,
  labels = markers,
  lines = F,
  facet = T,
  qq = T,
  pmax = NULL,
  models = c("GLM", "MLM", "CMLM", "MLMM", "SUPER", "FarmCPU", "Blink"),
  colors1 = c("darkgreen", "darkgoldenrod3", "darkgreen", "darkgoldenrod3",
    "darkgreen", "darkgoldenrod3", "darkgreen"),
  colors2 = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4",
    "darkgoldenrod2")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{lines}{Logical, value of whether or not to include vertical lines with markers.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot.}

\item{qq}{Logical, whether or not to add a QQ plot}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{colors1}{Colors for each chromosome.}

\item{colors2}{Colors for each model.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot.
}
