% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan.R
\name{gg_Manhattan}
\alias{gg_Manhattan}
\title{gg_Manhattan}
\usage{
gg_Manhattan(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  chrom = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  facet = F,
  addQQ = T,
  pmax = NULL,
  models = c("MLM", "FarmCPU", "BLINK", "MLMM", "GLM", "CMLM", "SUPER"),
  model.colors = c("darkgreen", "darkorange3", "steelblue", "darkred", "darkorchid4",
    "burlywood4", "darkseagreen4"),
  highlight.sig = F,
  sig.color = "darkred",
  chrom.colors = rep(c("darkgreen", "darkgoldenrod3"), 30),
  chrom.unit = "100 Mbp",
  legend.rows = 1,
  plotHBPvalues = F,
  skyline = "Kansas"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{chrom}{Chromosomes to plot. Use if you want to plot a single chromosome.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot. Default is `facet = F`.}

\item{addQQ}{Logical, whether or not to add a QQ plot.}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model. Used if `facet = F`.}

\item{highlight.sig}{Logical, whether or not to highlight significant associations with a black circle. Used if `facet = F`.}

\item{sig.color}{Color for significant assoctiations. Used if `facet = T`.}

\item{chrom.colors}{Colors for each chromosome. Used if `facet = T`.}

\item{chrom.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{legend.rows}{Number of rows for the legend.}

\item{plotHBPvalues}{Logical, if TRUE, H.B.P.values be uses.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If NULL, it will use the highest P.value.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot.
}
