% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan.R
\name{gg_Manhattan}
\alias{gg_Manhattan}
\title{gg_Manhattan}
\usage{
gg_Manhattan(
  folder,
  trait,
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = F,
  markers = NULL,
  labels = markers,
  facet = F,
  addQQ = T,
  pmax = NULL,
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM"),
  model.colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkorchid4"),
  chrom.colors = rep(c("darkgreen", "darkgoldenrod3"), 30),
  chrom.unit = "100 Mbp",
  legend.rows = 1
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{facet}{Logical, whether or not to produce a facetted or multi-model plot.}

\item{addQQ}{Logical, whether or not to add a QQ plot}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{model.colors}{Colors for each model. Used if `facet = F`.}

\item{chrom.colors}{Colors for each chromosome. Used if `facet = T`.}

\item{chrom.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{legend.rows}{Number of rows for the legend.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot.
}
