% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan_xModels.R
\name{gg_Manhattan_xModels}
\alias{gg_Manhattan_xModels}
\title{gg_Manhattan_xModels}
\usage{
gg_Manhattan_xModels(
  folder = "GWAS_Results/",
  traits = list_Traits(folder)[1:2],
  title = NULL,
  threshold = NULL,
  sug.threshold = NULL,
  chr = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  legend = F,
  legend.rows = 1,
  legend.box = "horizontal",
  point.sizes = c(0.3, 1, 0.75),
  addQQ = T,
  pmax = NULL,
  pmin = 0,
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM", "CMLM", "SUPER"),
  trait.colors = c("chartreuse4", "firebrick", "steelblue3", "maroon3", "purple3",
    "darkgoldenrod4", "tomato3", "aquamarine4", "deeppink3"),
  chr.unit = "100 Mbp",
  plotHBPvalues = F,
  skyline = NULL
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{traits}{The traits to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{chr}{Chromosomes to plot. Use if you want to plot a single chromosome.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{legend}{Logical, whether or not to add a legend.}

\item{legend.rows}{Number of rows for the legend.}

\item{legend.box}{Alignment of the legend. Default is "horizontal", but it can be changed to "vertical".}

\item{point.sizes}{Sizes for the points. c("Not Sig", "Sig", "Sug").}

\item{addQQ}{Logical, whether or not to add a QQ plot.}

\item{pmax}{A max value for the y-axis.}

\item{pmin}{A min Value for plotting. Markers with lower values will be removed.}

\item{models}{Model to read.}

\item{trait.colors}{Colors for each trait.}

\item{chr.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If left NULL, it will use the highest P.value.}
}
\value{
A manhattan plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/07_gg_Manhattan_xModels.html}{Create manhattan plots from GAPIT GWAS results for multiple traits and facets them by Model.}
}
