% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Manhattan_Zoom.R
\name{gg_Manhattan_Zoom}
\alias{gg_Manhattan_Zoom}
\title{gg_Manhattan_Zoom}
\usage{
gg_Manhattan_Zoom(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  chr,
  start,
  end,
  title = trait,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = "red",
  models = c("MLM", "FarmCPU", "BLINK", "MLMM", "GLM", "CMLM", "SUPER")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{chr}{Chromosome to plot.}

\item{start}{Start position on chromosome.}

\item{end}{End position on chromosome.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{models}{Models to read.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot zoomed in to a particular region.
}
