% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker.R
\name{gg_Marker}
\alias{gg_Marker}
\title{gg_Marker}
\usage{
gg_Marker(
  myG,
  myY,
  trait,
  marker,
  marker2 = NULL,
  type = "bar",
  points = T,
  colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4", "darkslategray",
    "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{myG}{GWAS genotype object. Note:}

\item{myY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{marker}{Marker to plot.}

\item{marker2}{Second marker to plot.}

\item{type}{Type of plot to make.}

\item{points}{Logical, whether or not to plot points}

\item{colors}{Color palette.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
