% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Bar.R
\name{gg_Marker_Bar}
\alias{gg_Marker_Bar}
\title{gg_Marker_Bar}
\usage{
gg_Marker_Bar(
  xG,
  xY,
  traits,
  markers,
  marker.colors = gwaspr_Colors,
  remove.hets = T,
  plot.histogram = T,
  plot.density = T,
  plot.counts = T,
  ncol = NULL,
  line.color = F,
  title = NULL,
  subtitle = paste(markers, collapse = "\\n"),
  asfactor = F
)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{traits}{Traits to plot.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}

\item{remove.hets}{Logical, Whether to remove hets or not. advised if plotting multiple markers.}

\item{plot.histogram}{Logical, if true will plot histogram bars.}

\item{plot.density}{Logical, if true will plot density bands.}

\item{plot.counts}{Logical, if true will make a plot of counts, if false will make a density plot.}

\item{ncol}{Number of columns for facetting when plotting multiple traits.}

\item{line.color}{Color of the border lines.}

\item{title}{Title for the plot.}

\item{subtitle}{Subtile for the plot. Defaults to the list of markers.}

\item{asfactor}{Logical, whether or not to plot as a factor or numeric.}
}
\value{
Marker plot.
}
\description{
\href{https://derekmichaelwright.github.io/gwaspr/articles/gg_Marker.html}{Creates a marker plot with myG and myY objects.}
}
