% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Bar.R
\name{gg_Marker_Bar}
\alias{gg_Marker_Bar}
\title{gg_Marker}
\usage{
gg_Marker_Bar(
  myG,
  myY,
  trait,
  markers,
  colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4", "darkslategray",
    "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{myG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{myY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{markers}{Markers to plot.}

\item{colors}{Color palette.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
