% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Bar.R
\name{gg_Marker_Bar}
\alias{gg_Marker_Bar}
\title{gg_Marker_Bar}
\usage{
gg_Marker_Bar(
  xG,
  xY,
  traits,
  markers,
  marker.colors = gwaspr_Colors,
  plot.histogram = T,
  plot.density = T,
  plot.counts = F,
  myncol = NULL,
  line.color = F
)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{traits}{Traits to plot.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}

\item{plot.histogram}{Logical, if true will plot histogram bars.}

\item{plot.density}{Logical, if true will plot density bands.}

\item{plot.counts}{Logical, if true will make a plot of counts, if false will make a density plot.}

\item{myncol}{Number of columns for facetting when plotting multiple traits.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
