% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Box.R
\name{gg_Marker_Box}
\alias{gg_Marker_Box}
\title{gg_Marker_Box}
\usage{
gg_Marker_Box(
  xG,
  xY,
  traits,
  markers,
  marker.colors = gwaspr_Colors,
  plot.violin = T,
  plot.box = T,
  plot.points = T,
  box.width = 0.1,
  point.size = 1,
  myncol = NULL,
  title = NULL,
  legend.rows = 1,
  subtitle = paste(markers, collapse = "\\n"),
  yLab = traits,
  xCV = NULL,
  cv.colors = NULL,
  cv.label = NULL
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{traits}{Traits to plot.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}

\item{plot.violin}{Logical, whether or not to plot violin.}

\item{plot.box}{Logical, whether or not to plot the boxplots.}

\item{plot.points}{Logical, whether or not to plot points.}

\item{box.width}{Width for the boxplot.}

\item{point.size}{Size for the points.}

\item{myncol}{Number of columns for facetting when plotting multiple traits.}

\item{title}{Title for the plot.}

\item{legend.rows}{Number of rows for the legend.}

\item{subtitle}{Subtitle for the plot. Defaults to the list of markers.}

\item{yLab}{Label for the y-axis.}

\item{xCV}{Covariate data.}

\item{cv.colors}{Covariate colors.}

\item{cv.label}{Label for the covariate.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
