% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Box.R
\name{gg_Marker_Box}
\alias{gg_Marker_Box}
\title{gg_Marker}
\usage{
gg_Marker_Box(
  myG,
  myY,
  trait,
  markers,
  points = T,
  colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4", "darkslategray",
    "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{myG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{myY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{markers}{Markers to plot.}

\item{points}{Logical, whether or not to plot points}

\item{colors}{Color palette.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
