% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Box.R
\name{gg_Marker_Box}
\alias{gg_Marker_Box}
\title{gg_Marker}
\usage{
gg_Marker_Box(
  xG,
  xY,
  traits,
  myncol = NULL,
  markers,
  box.width = 0.1,
  points = T,
  violin = T,
  marker.colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4",
    "darkslategray", "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{xG}{GWAS genotype object. Note: needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{traits}{Traits to plot.}

\item{markers}{Markers to plot.}

\item{box.width}{width for the boxplot.}

\item{points}{Logical, whether or not to plot points.}

\item{violin}{Logical, whether or not to plot violin.}

\item{marker.colors}{Color palette.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
