% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Pie.R
\name{gg_Marker_Pie}
\alias{gg_Marker_Pie}
\title{gg_Marker_Pie}
\usage{
gg_Marker_Pie(
  xG,
  xY,
  trait,
  trait.label = trait,
  markers,
  marker.colors = gwaspr_Colors,
  title = NULL,
  subtitle = paste(markers, collapse = "\\n")
)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{trait.label}{Label for the Trait.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}

\item{title}{Title for the plot.}

\item{subtitle}{Subtitle for the plot. Defaults to the list of markers.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
