% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Marker_Pie.R
\name{gg_Marker_Pie}
\alias{gg_Marker_Pie}
\title{gg_Marker}
\usage{
gg_Marker_Pie(
  xG,
  xY,
  trait,
  myncol = NULL,
  markers,
  marker.colors = c("darkgreen", "darkgoldenrod3", "darkred", "steelblue4",
    "darkslategray", "maroon4", "purple4", "darkblue")
)
}
\arguments{
\item{xG}{GWAS genotype object. Note:  needs to be in hapmap format.}

\item{xY}{GWAS phenotype object.}

\item{trait}{Trait to plot.}

\item{markers}{Markers to plot.}

\item{marker.colors}{Color palette.}
}
\value{
Marker plot.
}
\description{
Creates a marker plot with myG and myY objects.
}
