% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_myG_Details.R
\name{gg_myG_Details}
\alias{gg_myG_Details}
\title{gg_myG_Details}
\usage{
gg_myG_Details(filname, myPrefix = "")
}
\arguments{
\item{myPrefix}{Prefix for file names.}

\item{filename}{GWAS genotype object. Note: needs to be in hapmap format.}
}
\value{
2 .csv files with marker details & 3 Marker plots.
}
\description{
Analyses your genotype data and outputs 2 summary files with marker details and 3 plots.
}
