% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_NYCvsKansas.R
\name{gg_NYCvsKansas}
\alias{gg_NYCvsKansas}
\title{gg_NYCvsKansas}
\usage{
gg_NYCvsKansas(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  title = trait,
  threshold = NULL,
  chr.colors = rep(c("darkgreen", "darkgoldenrod3"), 30),
  chr.unit = "100 Mbp"
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{chr.colors}{Colors for each chromosome.}

\item{chr.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot comparing the NYC and Kansas results.
}
