% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_NYCvsKansas.R
\name{gg_NYCvsKansas}
\alias{gg_NYCvsKansas}
\title{gg_NYCvsKansas}
\usage{
gg_NYCvsKansas(
  folder = "GWAS_Results/",
  trait = list_Traits(folder)[1],
  title = trait,
  threshold = NULL,
  sug.threshold = NULL,
  chrom = NULL,
  markers = NULL,
  labels = markers,
  vlines = markers,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  addQQ = T,
  pmax = NULL,
  models = c("FarmCPU", "BLINK"),
  sig.col = "darkred",
  chrom.colors = rep(c("darkgreen", "darkgoldenrod3"), 30),
  chrom.unit = "100 Mbp",
  legend.rows = 1,
  plotHBPvalues = F
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{threshold}{Significant Threshold.}

\item{sug.threshold}{Suggested threshold.}

\item{chrom}{Chromosomes to plot. Use if you want to plot a single chromosome.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{vlines}{Markers which will be used as a location for a vertical lines.}

\item{vline.colors}{colors for each vertical line.}

\item{vline.types}{lty for each vertical line.}

\item{vline.legend}{Logical, whether or not to add a legend for the vlines.}

\item{addQQ}{Logical, whether or not to add a QQ plot}

\item{pmax}{A max value for the y-axis.}

\item{models}{Models to read.}

\item{sig.col}{Color for significant assoctiations. Used if `facet = T`.}

\item{chrom.colors}{Colors for each chromosome. Used if `facet = T`.}

\item{chrom.unit}{Unit for the x-axis. Can be one of c("kbp","100 kbp","Mbp","100 Mbp","Gbp").}

\item{legend.rows}{Number of rows for the legend.}

\item{plotHBPvalues}{Logical, if TRUE, H.B.P.Values be uses.}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot comparing the NYC and Kansas results.
}
