% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_QTL_Summary.R
\name{gg_QTL_Summary}
\alias{gg_QTL_Summary}
\title{gg_QTL_Summary}
\usage{
gg_QTL_Summary(
  xx = myQTLs,
  myG = myG,
  title = "Summary of QTL Results",
  caption = "derek was here"
)
}
\arguments{
\item{xx}{Table of QTL results.}

\item{myG}{genotype file}

\item{groups}{Grouping for the traits. Should be equal length to `traits`.}
}
\value{
A QTL summary plot.
}
\description{
Creates a summary QTL plot of significant associations.
Note: this function requires the GWAS results files to be ordered
}
