% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_QTL_Summary_Groups.R
\name{gg_QTL_Summary_Groups}
\alias{gg_QTL_Summary_Groups}
\title{gg_QTL_Summary_Groups}
\usage{
gg_QTL_Summary_Groups(
  myQ,
  myG,
  title = "Summary of QTL Results",
  yGroup,
  yLab = NULL,
  facetGroup,
  colorGroup,
  colorName = NULL,
  fillColors = gwaspr_Colors,
  xLab = "Pos",
  facetLab = "lg"
)
}
\arguments{
\item{myQ}{Table of QTL results. Must contain the following 5 columns c("Trait","Chr","Pos","Pos_lo","Pos_hi","lod"). then it must also include grouping columns set by the user for \code{yGroup},\code{facetGroup} & \code{colorGroup}.}

\item{myG}{Genetic map. Must contain the following 3 columns c("Marker","Chr","Pos").}

\item{title}{Custom title for the plot.}

\item{yGroup}{Name of column for y axis grouping.}

\item{yLab}{Custom y-axis label.}

\item{facetGroup}{Name of column for facet grouping.}

\item{colorGroup}{Name of column for color grouping.}

\item{colorName}{Title for the color legend.}

\item{fillColors}{Color for filling points.}

\item{xLab}{Custom x-axis label.}

\item{facetLab}{Custom label for facets.}
}
\value{
A QTL summary plot.
}
\description{
Creates a summary QTL plot of significant associations.
}
