% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_Volcano.R
\name{gg_Volcano}
\alias{gg_Volcano}
\title{gg_Volcano}
\usage{
gg_Volcano(
  folder,
  trait,
  title = trait,
  markers = NULL,
  labels = markers,
  models = c("GLM", "MLM", "MLMM", "CMLM", "SUPER", "FarmCPU", "BLINK")
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{trait}{The trait to read.}

\item{title}{A title for the plot.}

\item{markers}{Markers to be labelled.}

\item{labels}{Labels to be used for markers.}

\item{models}{Models to read.}
}
\value{
A volcano plot.
}
\description{
Creates a volcano plot.
}
