% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ordered.R
\name{is_ordered}
\alias{is_ordered}
\title{is_ordered}
\usage{
is_ordered(folder = "GWAS_Results/")
}
\arguments{
\item{folder}{Folder containing GWAS results.}
}
\value{
A table of which models have been run on each trait.
}
\description{
Create a table of which GWAS results files are ordered by p.value.
}
