% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ran.R
\name{is_ran}
\alias{is_ran}
\title{is_ran}
\usage{
is_ran(folder = "GWAS_Results/")
}
\arguments{
\item{folder}{Folder containing GWAS results.}
}
\value{
A table of which models have been run on each trait.
}
\description{
Create a table of which models have been run on each trait.
}
