% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Top_Markers.R
\name{list_Top_Markers}
\alias{list_Top_Markers}
\title{list_Top_Markers}
\usage{
list_Top_Markers(x, chroms = 1:7, n = 1, threshold = NULL)
}
\arguments{
\item{x}{data.}

\item{chroms}{Chromosomes to include.}

\item{n}{Number per chromosome.}

\item{threshold_filter}{Logical, filters non-significant results.}
}
\value{
Table of top results.
}
\description{
Finds the markers with the highest association on each chromosome.
}
