% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Top_Markers.R
\name{list_Top_Markers}
\alias{list_Top_Markers}
\title{list_Top_Markers}
\usage{
list_Top_Markers(
  trait,
  model,
  folder = NULL,
  threshold = NULL,
  chroms = 1:50,
  n = 1
)
}
\arguments{
\item{trait}{GWAS trait.}

\item{threshold}{filters results with -log10(p) below threshold.}

\item{chroms}{Chromosomes to include.}

\item{n}{Number per chromosome.}

\item{method}{GWAS method.}
}
\value{
Table of top results.
}
\description{
Finds the markers with the highest association on each chromosome.
}
