% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GWAS_Results.R
\name{table_GWAS_Results}
\alias{table_GWAS_Results}
\title{table_GWAS_Results}
\usage{
table_GWAS_Results(
  folder,
  file,
  threshold = 6,
  sug.threshold = NULL,
  rowread = 2000
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{file}{The file to read.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{rowread}{Number of rows to read.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results.
}
