% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GWAS_Results.R
\name{table_GWAS_Results}
\alias{table_GWAS_Results}
\title{table_GWAS_Results}
\usage{
table_GWAS_Results(
  folder = "GWAS_Results/",
  fnames = list_Result_Files(folder),
  nrowstoread = 1000,
  threshold = 6,
  sug.threshold = NULL,
  skyline = NULL
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{fnames}{The files to read.}

\item{nrowstoread}{Number of rows to read.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{skyline}{Which skyline type to use. Can be "NYC" or "Kansas". If left NULL, it will use the highest P.value.}

\item{useHBPvalues}{Logical, if TRUE, H.B.P.Values will be uses.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results.
}
