% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GWAS_Results.R
\name{table_GWAS_Results}
\alias{table_GWAS_Results}
\title{table_GWAS_Results}
\usage{
table_GWAS_Results(
  folder = "GWAS_Results/",
  files = list_Result_Files(folder),
  nrowstoread = 1000,
  threshold = 6,
  sug.threshold = NULL,
  useHBPvalues = F
)
}
\arguments{
\item{folder}{Folder containing GWAS results.}

\item{files}{The files to read.}

\item{nrowstoread}{Number of rows to read.}

\item{threshold}{Significant threshold.}

\item{sug.threshold}{Suggestive threshold.}

\item{useHBPvalues}{Logical, should H.B.P.Values be uses.}
}
\value{
A table of significant GWAS results.
}
\description{
Create a table of significant GWAS results.
}
