% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gwaspr_Pie.R
\name{theme_gwaspr_pie}
\alias{theme_gwaspr_pie}
\title{theme_gwaspr_pie.}
\usage{
theme_gwaspr_pie(
  x,
  bgFill = "white",
  lineColor = "grey95",
  linesize = 0.75,
  stripFill = "white",
  ...
)
}
\arguments{
\item{x}{ggplot}
}
\value{
ggplot with agData theme for pie graphs
}
\description{
ggplot theme.
}
